Predict Method for Wordfish

Predicts positions of new documents using a fitted Wordfish model

# S3 method for wordfish
predict(object, newdata = NULL, se.fit = FALSE,
  interval = c("none", "confidence"), level = 0.95, ...)

Arguments

object
A fitted wordfish model

newdata
An optional data frame or object of class wfm in which to look for word counts to predict document ideal points which to predict. If omitted, the fitted values are used.

se.fit
A switch indicating if standard errors are required.

interval
Type of interval calculation

level
Tolerance/confidence level

...
further arguments passed to or from other methods.

Value

predict.wordfish produces a vector of predictions or a matrix of predictions and bounds with column names `fit' and `se.fit', and with `lwr', and `upr' if `interval' is also set.

Details

Standard errors for document positions are generated by numerically inverting the relevant Hessians from the profile likelihood of the multinomial form of the model.

See also

wordfish